requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
google-generativeai==0.8.3
google-api-core==2.19.2

aiogram==3.5.0
faster-whisper==1.0.3
httpx==0.27.2
requests==2.32.3
python-telegram-bot==13.15
urllib3==1.26.18

APScheduler==3.6.3
gspread==6.0.2
google-auth==2.35.0
